backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.11
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.25.1
pydantic>=2.4.2
python-multipart>=0.0.6
PyJWT>=2.8.0
